[Projects/Test] Move workflow store:// tests into TestKFP
Address Liran's review on the workflow store:// system tests:
- Move both parametrized tests (engine=kfp / engine=remote) into the
  existing TestKFP class in tests/system/runtimes/test_kfp.py.
- Inline the single-use pipeline-body string into the helper that writes it.
- Per-test try/finally for tempfile cleanup, so existing TestKFP tests
  are not affected (no class-level setup/teardown changes).
- Delete the now-empty tests/system/projects/test_workflow_store_uri.py.
